+++ lib/event.dart
@@ -248,3 +248,5 @@ part 'src/event/adapter/touch_cancel_handler_adapter.dart';
 part 'src/event/adapter/touch_end_handler_adapter.dart';
 part 'src/event/adapter/touch_move_handler_adapter.dart';
 part 'src/event/adapter/touch_start_handler_adapter.dart';
+
+part 'src/event/adapter/all_mouse_handlers_adapter.dart';