+++ test/run.sh
@@ -6,18 +6,18 @@ set -e
 # Type Analysis
 ###############
 
-echo
-echo "dart_analyzer lib/*.dart"
+#echo
+#echo "dart_analyzer lib/*.dart"
 
-results=`dart_analyzer lib/*.dart 2>&1`
+#results=`dart_analyzer lib/*.dart 2>&1`
 
-echo "$results"
+#echo "$results"
 
-if [ -n "$results" ]; then
-    exit 1
-else
-    echo "Passed analysis."
-fi
+#if [ -n "$results" ]; then
+#    exit 1
+#else
+#    echo "Passed analysis."
+#fi
 
 ############
 # Unit Tests